@@ -1,19 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-    </PropertyGroup>
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
 
-    <ItemGroup>
-        <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.2" />
-        <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.10.0" />
-        <PackageReference Include="NUnit" Version="3.14.0" />
-        <PackageReference Include="NUnit3TestAdapter" Version="4.6.0" />
-        <PackageReference Include="Shuttle.Core.Data.Logging" Version="17.0.0" />
-        <PackageReference Include="Shuttle.Esb.Tests" Version="15.0.0" />
-    </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="5.2.2" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
+    <PackageReference Include="NUnit" Version="4.2.2" />
+    <PackageReference Include="NUnit3TestAdapter" Version="4.6.0" />
+    <PackageReference Include="Shuttle.Core.Data.Logging" Version="20.0.0" />
+    <PackageReference Include="Shuttle.Esb.Tests" Version="20.0.0" />
+  </ItemGroup>
 
-    <ItemGroup>
-        <ProjectReference Include="..\Shuttle.Esb.Sql.Queue\Shuttle.Esb.Sql.Queue.csproj" />
-    </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Shuttle.Esb.Sql.Queue\Shuttle.Esb.Sql.Queue.csproj" />
+  </ItemGroup>
 </Project>
