@@ -2,20 +2,13 @@ using System.Threading.Tasks;
 using NUnit.Framework;
 using Shuttle.Esb.Tests;
 
-namespace Shuttle.Esb.Sql.Queue.Tests
-{
-	public class SqlPipelineExceptionHandlingFixture : PipelineExceptionFixture
-	{
-		[Test]
-		public void Should_be_able_to_handle_exceptions_in_receive_stage_of_receive_pipeline()
-		{
-			TestExceptionHandling(SqlConfiguration.GetServiceCollection(), "sql://shuttle/{0}");
-		}
+namespace Shuttle.Esb.Sql.Queue.Tests;
 
-		[Test]
-		public async Task Should_be_able_to_handle_exceptions_in_receive_stage_of_receive_pipeline_async()
-		{
-			await TestExceptionHandlingAsync(SqlConfiguration.GetServiceCollection(), "sql://shuttle/{0}");
-		}
-	}
+public class SqlPipelineExceptionHandlingFixture : PipelineExceptionFixture
+{
+    [Test]
+    public async Task Should_be_able_to_handle_exceptions_in_receive_stage_of_receive_pipeline_async()
+    {
+        await TestExceptionHandlingAsync(SqlConfiguration.GetServiceCollection(), "sql://shuttle/{0}");
+    }
 }
