@@ -1,17 +1,15 @@
-namespace Shuttle.Esb.Sql.Queue
+namespace Shuttle.Esb.Sql.Queue;
+
+public class Script
 {
-	public class Script
-	{
-		public static readonly string QueueCount = "QueueCount";
-		public static readonly string QueueCreate = "QueueCreate";
-		public static readonly string QueueDequeue = "QueueDequeue";
-		public static readonly string QueueDequeueId = "QueueDequeueId";
-		public static readonly string QueueDrop = "QueueDrop";
-		public static readonly string QueueEnqueue = "QueueEnqueue";
-		public static readonly string QueueExists = "QueueExists";
-		public static readonly string QueuePurge = "QueuePurge";
-		public static readonly string QueueRemove = "QueueRemove";
-		public static readonly string QueueRead = "QueueRead";
-		public static readonly string QueueRelease = "QueueRelease";
-	}
+    public static readonly string QueueCount = "QueueCount";
+    public static readonly string QueueCreate = "QueueCreate";
+    public static readonly string QueueDequeue = "QueueDequeue";
+    public static readonly string QueueDequeueId = "QueueDequeueId";
+    public static readonly string QueueDrop = "QueueDrop";
+    public static readonly string QueueEnqueue = "QueueEnqueue";
+    public static readonly string QueueExists = "QueueExists";
+    public static readonly string QueuePurge = "QueuePurge";
+    public static readonly string QueueRemove = "QueueRemove";
+    public static readonly string QueueRelease = "QueueRelease";
 }
