@@ -1,22 +1,11 @@
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-#if NETFRAMEWORK
-[assembly: AssemblyTitle(".NET Framework")]
-#endif
-
-#if NETCOREAPP
-[assembly: AssemblyTitle(".NET Core")]
-#endif
-
-#if NETSTANDARD
-[assembly: AssemblyTitle(".NET Standard")]
-#endif
-
-[assembly: AssemblyVersion("16.0.1.0")]
+[assembly: AssemblyTitle(".NET Unified Platform")]
+[assembly: AssemblyVersion("20.0.0.0")]
 [assembly: AssemblyCopyright("Copyright (c) 2025, Eben Roux")]
 [assembly: AssemblyProduct("Shuttle.Esb.Sql.Queue")]
 [assembly: AssemblyCompany("Eben Roux")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyInformationalVersion("16.0.1")]
+[assembly: AssemblyInformationalVersion("20.0.0")]
 [assembly: ComVisible(false)]
